\name{TsayData}

\alias{TsayData}

\alias{dciti.txt}
\alias{dcscointc.txt}
\alias{dcscoy99ln.txt}
\alias{dgnp82.txt}
\alias{dibmln.txt}
\alias{dibmln98.txt}
\alias{dibmln99.txt}
\alias{dintc.txt}
\alias{dmmm.txt}
\alias{dmsft.txt}
\alias{dvwew.txt}
\alias{exchperc.txt}
\alias{hkja.txt}
\alias{ibm.txt.txt}
\alias{ibm1to5dur.txt}
\alias{ibm91ads.txt}
\alias{ibm91adsx.txt}
\alias{ibm9912tp.txt}
\alias{ibmdurad.txt}
\alias{ibmy98.txt}
\alias{jnj.dat}
\alias{m3m4699.txt}
\alias{m5cln.txt}
\alias{mew.txt}
\alias{mgeln.txt}
\alias{mgs.txt}
\alias{mgs1n3.txt}
\alias{mibmln.txt}
\alias{mibmln99.txt}
\alias{mibmsp6299.txt}
\alias{mibmspln.txt}
\alias{mibmsplnsu.txt}
\alias{mintc.txt}
\alias{mmmm.txt}
\alias{msp6299.txt}
\alias{mvw.txt}
\alias{qgnp.txt}
\alias{qunemrate.txt}
\alias{sp500.txt}
\alias{sp5may.txt}
\alias{wgs1yr.txt}
\alias{wgs3c.txt}
\alias{wgs3n1c.txt}
\alias{wgs3yr.txt}
\alias{wtb3ms.txt}
\alias{wtb6ms.txt}
	 		

\title{Ruey Tsay's Data Sets}


\description{

	A collection and description of data sets 
	from the book 'Analysis of Financial Time
	Series" written by Ruey S. Tsay.
	\cr
	
	The data sets are:
		
	Chapter 1, Financial Time Series and Their Characteristics
	
	\tabular{ll}{
	\code{dibmln.txt} \tab Daily log returns of IBM (62/7/3 to 97/12), \cr  
	\code{dvwew.txt} \tab Daily simple returns of valueweighted and equalweighted indexes, \cr 
	\code{dintc.txt} \tab Daily simple returns of Intel stock, \cr   
	\code{dmmm.txt} \tab Daily simple returns of 3M stock, \cr 
	\code{dmsft.txt} \tab Daily simple returns of Microsoft stock, \cr 
	\code{dciti.txt} \tab Daily simple returns of Citigroup stock, \cr 
	\code{mbnd.txt} \tab Monthly bond returns (30 yrs, 20 yrs, ..., 1 yr), \cr 
	\code{mgs.txt} \tab Monthly Treasury rates (10 yrs, 5 yrs, ..., 1 yr), \cr 
	\code{wtb3ms.txt} \tab 3M Weekly Treasury Bill rates, \cr 
	\code{wtb6ms.txt} \tab 6M Weekly Treasury Bill rates. }
	
	
	Chapter 2, Linear Time Series Analysis and Its Applications
	
	\tabular{ll}{
	\code{qgnp.txt} \tab US quarterly growth rates of GNP, \cr 
	\code{mvw.txt} \tab Monthly valueweighted index returns, \cr 
	\code{mew.txt} \tab Monthly equalweighted index returns, \cr 
	\code{m3m4699.txt} \tab Monthly log returns of 3M stock, \cr 
	\code{jnj.txt} \tab Quarterly earnings per share of Johnson and Johnson, \cr 
	\code{wgs1yr.txt} \tab Weekly US Treasury 1y constant maturity rates, \cr  
	\code{wgs3yr.txt} \tab Weekly US Treasury 3y constant maturity rates. } 
	
	    
	Chapter 3, Conditional Heteroscedastic Models
	
	\tabular{ll}{
	\code{mintc.txt} \tab Monthly simple returns of Intel stock, \cr 
	\code{exchperc.txt} \tab 10m log returns of FX (MarkUS), \cr 
	\code{sp500.txt} \tab Excess returns of S&P500, \cr 
	\code{mibmln.txt} \tab Monthly log returns of IBM stock, \cr 
	\code{dhwp3dx8099.txt} \tab Daily log returns of SP500 index, \cr 
	\code{mibmspln.txt} \tab Monthly log returns of IBM stock and SP500, \cr 
	\code{mibmsplnsu.txt} \tab Data set for Example 3.5. }
	
	    
	Chapter 4, Nonlinear Models and Their Applications
	
	\tabular{ll}{
	\code{mew.txt} \tab Monthly simple returns of equalweighted index, \cr 
	\code{dibmln99.txt} \tab Daily log returns of IBM stock, \cr 
	\code{mmmm.txt} \tab Monthly simple returns of 3M stock, \cr 
	\code{qgnp.txt} \tab Quarterly growth rates of US gnp, \cr 
	\code{mibmln99.txt} \tab Monthly log returns of IBM stock, \cr 
	\code{qunemrate.txt} \tab Quarterly unemployment rates. }
	
	
	Chapter 5, HighFrequency Data Analysis and Market Microstructure
	
	\tabular{ll}{
	\code{ibm.txt} \tab IBM transactions data 1990-11-01 1991-01-31, \cr   
	\code{ibm9912tp.txt} \tab IBM transactions data of December 1999, \cr   
	\code{ibmdurad.txt} \tab IBM Adjusted time durations between trades, \cr 
	\code{ibm1to5dur.txt} \tab IBM Adjusted durations for first 5 trading days, \cr 
	\code{ibm91ads.txt} \tab IBM Data for Example 5.2 - ADS file, \cr 
	\code{ibm91adsx.txt} \tab IBM explanatory variables as defined, \cr 
	\code{day15ori.txt} \tab IBM Transactions data 1990-11-21, original data, \cr 
	\code{day15.txt} \tab IBM Data for PCD models.}
	
	    
	Chapter 7, Extreme Values, Quantile Estimation, and Value at Risk
	
	\tabular{ll}{
	\code{dibmln98.txt} \tab Daily perdentage log returns of IBM stock, \cr 
	\code{dintc7297.txt} \tab Daily log returns of Intel stock, Example 7.4, \cr 
	\code{ibmln98wm.txt} \tab Meancorrected daily log returns of IBM, \cr 
	\code{ibml25x.txt} \tab The explanatory variables on page 294. }
	
	    
	Chapter 8, Multivariate Time Series Analysis and Its Applications
	
	\tabular{ll}{
	\code{mibmspln.txt} \tab Monthly log returns of IBM and SP 500, \cr 
	\code{mbnd.txt} \tab Monthly simple returns of bond indexes, \cr  
	\code{mgs1n3.txt} \tab Monthly US interest rates, Example 8.6, \cr  
	\code{sp5may.txt} \tab Log prices of SP500 index futures and shares, \cr 
	\code{m5cln.txt} \tab Monthly log returns of IBM, HWP, INTC, MER and MWD. }
	
	
	Chapter 9, Multivariate Volatility Models and Their Applications
	
	\tabular{ll}{
	\code{hkja.txt} \tab Daily log returns of HK and Japan market index, \cr 
	\code{mibmspln.txt} \tab Monthly log returns of IBM and SP 500, \cr 
	\code{dcscointc.txt} \tab Daily log returns of SP 500, Cisco and Intel stocks. } 
	
	    
	Chapter 10, Markov Chain Monte Carlo Methods with Applications
	
	\tabular{ll}{
	\code{wgs3n1c.txt} \tab Change series of weekly US interest rates, 3Y and 1Y, \cr 
	\code{wgs3c.txt} \tab Change series of weekly US 3yr interest rate, \cr 
	\code{msp6299.txt} \tab Monthly log returns of SP 500 index, \cr 
	\code{mibmsp6299.txt} \tab Monthly log returns of IBM stock and SP 500, \cr 
	\code{mgeln.txt} \tab Monthly log returns of GE stock. }
	   
}


\references{

Tsay, R.S. (2002);
	\emph{ Analysis of Financial Time Series},
	Wiley.
	
}


\examples{
## SOURCE("fBasics.36C-TsayData")

## *.txt - 
  
}


\keyword{datasets}

