\name{fapply}


\title{Applies Functions Over timeSeries Margins}

\alias{fapply}
\alias{apply}



\description{
    
    Applies Functions Over timeSeries Margins          
    
}


\usage{
fapply(x, from, to, FUN, \dots)
}


\arguments{
  
    \item{from, to}{
        starting date and end date, \code{to} must be after \code{from}.
        }
    \item{FUN}{
        the function to be applied.
        }
    \item{x}{
        an object of class \code{timeSeries}.
        } 
    \item{\dots}{
        arguments passed to other methods.
        }
       
}


\author{

    Diethelm Wuertz for the Rmetrics \R-port.
    
}


\examples{
## data - 

}


\keyword{chron}

