\name{fapply}


\title{Apply Function Over Time Series Margins}

\alias{fapply}
\alias{apply}



\description{
    
    Apply a functions over 'timeSeries' margins.          
    
}


\usage{
fapply(x, from, to, FUN, \dots)
}


\arguments{
  
    \item{from, to}{
        starting date and end date, \code{to} must be after \code{from}.
        }
    \item{FUN}{
        the function to be applied.
        }
    \item{x}{
        an object of class \code{timeSeries}.
        } 
    \item{\dots}{
        arguments passed to other methods.
        }
       
}


\examples{
## data - 

}


\keyword{chron}

