\name{fSeries-deprecated}

\title{Deprecated functions in fSeries package}

\alias{seriesPositions}
\alias{newPositions<-}

\description{
  
    \tabular{ll}{   
    \code{seriesPositions} \tab Extracts positions slot from a 'timeSeries', \cr
    \code{newPositions<-} \tab Modifies positions of a 'timeSeries' object, \cr
  }

}
  